@ Widen instance-id handles to long; adopt EntityId from 6000.3
Builds on the 6000.5 EntityId migration. Moves the #if gate to 6000.3,
where the int-based instance-ID APIs are already deprecated (CS0618) and
EntityId exists, so 6.3/6.4 stop emitting deprecation warnings. Uses a
three-way path because EntityId is int-backed on 6000.3/.4 (implicit
int conversions) and only 64-bit (ToULong/FromULong) on 6000.5+.

Surfaces handles as long end-to-end so a 6000.5 64-bit EntityId is never
truncated; narrower ids widen and narrow back exactly:
- Protocol DTOs: InstanceId, OpenedFromInstanceId, BlockedBy
- CLI: --id options (snapshot, ui click), ParseInstanceIds + eval preamble
- Package: wrappers, GetLongArgument, targetId/context paths, StageInfo
- Rebuilt bundled UnityCtl.Protocol.dll

Verified on 6000.3.15f1: compiles clean (no Safe Mode), deprecation
warnings gone, snapshot --id round-trips. dotnet build + 482 tests pass.
@

+++ UnityCtl.Cli/ScriptCommands.cs
@@ -613,15 +613,15 @@ public static class ScriptCommands
         return expression;
     }
 
-    internal static int[] ParseInstanceIds(string idArg)
+    internal static long[] ParseInstanceIds(string idArg)
     {
         var parts = idArg.Split(',');
-        var ids = new List<int>();
+        var ids = new List<long>();
         foreach (var part in parts)
         {
             var trimmed = part.Trim();
             if (trimmed.Length == 0) continue;
-            if (!int.TryParse(trimmed, out var id))
+            if (!long.TryParse(trimmed, out var id))
                 throw new ArgumentException($"Invalid instance ID: '{trimmed}' — must be an integer");
             ids.Add(id);
         }
@@ -630,7 +630,7 @@ public static class ScriptCommands
         return ids.ToArray();
     }
 
-    internal static string BuildInstanceIdPreamble(int[] ids)
+    internal static string BuildInstanceIdPreamble(long[] ids)
     {
         const string pad = "\n        ";
         var sb = new System.Text.StringBuilder();
